+++ src/ui/ui/pool.tpl
@@ -15,7 +15,7 @@
                                 <input type="text" name="name" class="form-control"
                                     placeholder="{$_L['Search_by_Name']}...">
                                 <div class="input-group-btn">
-                                    <button class="btn btn-success">{$_L['Search']}</button>
+                                    <button class="btn btn-success" type="submit">{$_L['Search']}</button>
                                 </div>
                             </div>
                         </form>
@@ -47,7 +47,7 @@
                                     <td align="center">
                                         <a href="{$_url}pool/edit/{$ds['id']}"
                                             class="btn btn-info btn-xs">{$_L['Edit']}</a>
-                                        <a href="{$_url}pool/delete/{$ds['id']}" id="{$ds['id']}"
+                                        <a href="{$_url}pool/delete/{$ds['id']}" id="{$ds['id']}" onclick="return confirm('{$_L['Delete']}?')"
                                             class="btn btn-danger btn-xs">{$_L['Delete']}</a>
                                     </td>
                                 </tr>
